RF: split ballot app into ballot_measure and election_day apps.
