+++ ballot_measure/admin.py
@@ -1,0 +1,3 @@
+from django.contrib import admin
+
+admin.autodiscover()
