+++ election_day/serializers.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from .models import ElectionDay
+
+
+class ElectionDaySerializer(serializers.ModelSerializer):
+
+    class Meta:
+        model = ElectionDay
